Refacto : move Beans from MasterpieceApplication to config/BeanConfig Class
